+++ src/karcarthy/run/context.clj
@@ -83,11 +83,11 @@
   ([event]
    (emit! (current-run-context) event))
   ([ctx event]
-   (let [event (merge {:karcarthy/type :event
+   (let [event (merge event
+                      {:karcarthy/type :event
                        :time-ms (System/currentTimeMillis)
                        :run-id (:run-id ctx)}
-                      (select-keys ctx [:agent-id :parent-id :depth])
-                      event)]
+                      (select-keys ctx [:agent-id :parent-id :depth]))]
      (when-not (contains? #{:model/text-delta :model/tool-call-delta
                             :model/stream-event}
                           (:type event))